added some user permissions to the time slots page

+++ ParishManager/Controllers/TimeSlotController.cs
@@ -20,22 +20,26 @@ namespace ParishManager.Controllers
         private readonly ITimeSlotCommitmentService _commitmentService;
         private readonly ITimeService _timeService;
         private readonly UserManager<User> _userManager;
+        private readonly IUserService _userService;
 
         public TimeSlotController(
             ITimeSlotService timeSlotService,
             ITimeSlotCommitmentService commitmentService,
             ITimeService timeService,
-            UserManager<User> userManager)
+            UserManager<User> userManager,
+            IUserService userService)
         {
             _timeSlotService = timeSlotService;
             _commitmentService = commitmentService;
             _timeService = timeService;
             _userManager = userManager;
+            _userService = userService;
         }
 
         public IActionResult Index(int parishId = 1, string alertMessageText = null)
         {
             var userId = _userManager.GetUserId(User);
+            var isAdmin = _userService.IsAdminForParish(userId, parishId);
 
             var timeSlots = _timeSlotService.GetTimeSlotsByParishId(userId, parishId)
                 .Select(x => new TimeSlotListItemViewModel()
@@ -51,7 +55,8 @@ namespace ParishManager.Controllers
             {
                 TimeSlots = timeSlots,
                 AlertMessage = alertMessageText,
-                ParishId = parishId
+                ParishId = parishId,
+                IsAdmin = isAdmin
             };
 
             return View(viewModel);
@@ -77,6 +82,14 @@ namespace ParishManager.Controllers
 
         public IActionResult Create(int parishId)
         {
+            var userId = _userManager.GetUserId(User);
+            var isAdmin = _userService.IsAdminForParish(userId, parishId);
+
+            if (!isAdmin)
+            {
+                return RedirectToAction("Index", new { parishId, alertMessageText = "Could not create time slot due to improper access rights." });
+            }
+
             var hoursList = GetHoursList();
 
             var viewModel = new TimeSlotCreateViewModel
@@ -91,6 +104,17 @@ namespace ParishManager.Controllers
         [HttpPost]
         public IActionResult Create(TimeSlotCreateViewModel model)
         {
+            var userId = _userManager.GetUserId(User);
+            var isAdmin = _userService.IsAdminForParish(userId, model.ParishId);
+
+            if (!isAdmin)
+            {
+                return RedirectToAction("Index", new {
+                    parishId = model.ParishId,
+                    alertMessageText = "Could not create time slot due to improper access rights."
+                });
+            }
+
             var timeSlot = new TimeSlotCreate()
             {
                 Day = model.Day,
@@ -103,12 +127,26 @@ namespace ParishManager.Controllers
 
             _timeSlotService.Create(timeSlot);
 
-            return RedirectToAction("Index", new { parishId = model.ParishId, alertMessageText = "New time slot successfully created!" });
+            return RedirectToAction("Index", new {
+                parishId = model.ParishId,
+                alertMessageText = "New time slot successfully created!"
+            });
         }
 
         public IActionResult Edit(int id)
         {
             var timeSlot = _timeSlotService.Get(id);
+
+            var userId = _userManager.GetUserId(User);
+            var isAdmin = _userService.IsAdminForParish(userId, timeSlot.Parish.Id);
+
+            if (!isAdmin)
+            {
+                return RedirectToAction("Index", new {
+                    parishId = timeSlot.Parish.Id,
+                    alertMessageText = "Could not edit time slot due to improper access rights."
+                });
+            }
 
             var committedUsers = _commitmentService
                 .GetCommitedUsersForTimeSlot(id)
@@ -135,6 +173,9 @@ namespace ParishManager.Controllers
         [HttpPost]
         public IActionResult Edit(TimeSlotEditModel model)
         {
+            // really need to check if this user is authenticated
+            // will require getting the time slot and then editing the time slot
+
             var timeSlot = new TimeSlotUpdate()
             {
                 Id = model.TimeSlotId,
@@ -161,6 +202,17 @@ namespace ParishManager.Controllers
         public IActionResult Delete(int id)
         {
             var timeSlot = _timeSlotService.Get(id);
+
+            var userId = _userManager.GetUserId(User);
+            var isAdmin = _userService.IsAdminForParish(userId, timeSlot.Parish.Id);
+
+            if (!isAdmin)
+            {
+                return RedirectToAction("Index", new {
+                    parishId = timeSlot.Parish.Id,
+                    alertMessageText = "Could not delete time slot due to improper access rights."
+                });
+            }
 
             _timeSlotService.Delete(id);
